@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,17 +24,17 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  dio: 3.0.9
-  toast: ^0.1.5
-  flutter_swiper : ^1.1.6
-  shared_preferences: ^2.0.7
-  flutter_webview_plugin: ^0.4.0
-  date_format: ^1.0.8
-  provider: ^6.0.0
-#  webview_flutter: ^0.3.10+3
-  local_auth: ^1.1.7
+  cupertino_icons: ^1.0.5
+  dio: ^4.0.6
+  toast: ^0.3.0
+  flutter_swiper_null_safety_flutter3: ^2.0.0
+  shared_preferences: ^2.0.15
+  flutter_inappwebview: ^5.4.3+7
+  date_format: ^2.0.6
+  provider: ^6.0.3
+  local_auth: ^2.1.1
   gesture_password_widget: ^2.0.0
+  url_launcher: ^6.1.5
 
 dev_dependencies:
   flutter_test:
