@@ -1,12 +1,12 @@
 class BannerEntity {
-	String imagePath;
-	int id;
-	int isVisible;
-	String title;
-	int type;
-	String url;
-	String desc;
-	int order;
+	String? imagePath;
+	int? id;
+	int? isVisible;
+	String? title;
+	int? type;
+	String? url;
+	String? desc;
+	int? order;
 
 	BannerEntity({this.imagePath, this.id, this.isVisible, this.title, this.type, this.url, this.desc, this.order});
 
