@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:local_auth/auth_strings.dart';
 import 'package:local_auth/local_auth.dart';
-import 'package:wan_android/constant/custom_arguments.dart';
 import 'package:wan_android/constant/http_code.dart';
 import 'package:wan_android/constant/widget_style.dart';
 import 'package:wan_android/entity/base_entity.dart';
@@ -146,33 +143,32 @@ class _LoginPageState extends State<LoginPage> {
     BaseEntity baseEntity =
         await HttpManager.getInstance().login(API.LOGIN_URL, params);
     if (HttpCode.SUCCESS == baseEntity.code) {
-      BaseData baseData = baseEntity.data;
-      int errorCode = baseData.errorCode;
+      BaseData? baseData = baseEntity.data;
+      int? errorCode = baseData?.errorCode;
       if (errorCode == HttpCode.ERROR_CODE_SUC) {
         print('登录成功');
         ToastUtil.showToast(context, '登录成功');
-        UserEntity userEntity = UserEntity.fromJson(baseData.data);
+        UserEntity userEntity = UserEntity.fromJson(baseData?.data);
         //账号和id保存在本地,跳转到首页
-        User().saveUsername(userEntity.username);
+        User().saveUsername(userEntity.username??'');
         //检测当前设备是否支持指纹
         bool isDeviceSupport = await auth.isDeviceSupported();
         if (isDeviceSupport) {
           //支持指纹或面部识别
           onShowDialog(context, "检测到您的设备支持指纹或扫脸识别，是否开启，以便下次快速登录？", () {
             //用户取消设置指纹
-            onShowDialog(context, '是否设置手势图案快速登录？', (){
+            onShowDialog(context, '是否设置手势图案快速登录？', () {
               //用户取消设置图案，关闭登录页
               Navigator.pop(context);
-            }, ()=>_setGesture(context));
+            }, () => _setGesture(context));
           }, () => _setFingerprint(context));
         } else {
           //不支持生物识别，提示设置手势识别
-          onShowDialog(context, '是否设置手势图案快速登录？', (){
+          onShowDialog(context, '是否设置手势图案快速登录？', () {
             //用户取消设置图案，关闭登录页
             Navigator.pop(context);
-          }, ()=>_setGesture(context));
+          }, () => _setGesture(context));
         }
-
       } else {
         print('登录失败');
         ToastUtil.showToast(context, '登录失败:$errorCode');
@@ -196,41 +192,39 @@ class _LoginPageState extends State<LoginPage> {
     Navigator.pushNamed(context, '/GestureUnlockPage');
   }
 
-  void _setFingerprint(context) async{
+  void _setFingerprint(context) async {
     //设置指纹登录
     //弹框汉化
-    try {
-      const androidAuth = const AndroidAuthMessages(
-            cancelButton: '取消',
-            goToSettingsButton: '去设置',
-            goToSettingsDescription: '请设置指纹.',
-            biometricHint: '身份校验',
-            signInTitle: '指纹验证',
-          );
-      bool authenticated = await auth.authenticate(
-              localizedReason:
-              '扫描指纹进行身份识别',
-              useErrorDialogs: false,
-              stickyAuth: true,
-              androidAuthStrings: androidAuth,
-              biometricOnly: true);
-      if(authenticated){
-        //识别成功，记录状态，关闭登录页
-        User().saveFingerprint(true);
-        Navigator.pop(context);
-      }else{
-        //识别失败，提示设置手势
-        onShowDialog(context, '是否设置手势图案快速登录？', (){
-          //用户取消设置图案，关闭登录页
-          Navigator.pop(context);
-        }, ()=>_setGesture(context));
-      }
-    } on PlatformException catch (e) {
-      print(e);
-    }
+    // try {
+    //   const androidAuth = const AndroidAuthMessages(
+    //         cancelButton: '取消',
+    //         goToSettingsButton: '去设置',
+    //         goToSettingsDescription: '请设置指纹.',
+    //         biometricHint: '身份校验',
+    //         signInTitle: '指纹验证',
+    //       );
+    //   bool authenticated = await auth.authenticate(
+    //           localizedReason:
+    //           '扫描指纹进行身份识别',
+    //           useErrorDialogs: false,
+    //           stickyAuth: true,
+    //           androidAuthStrings: androidAuth,
+    //           biometricOnly: true);
+    //   if(authenticated){
+    //     //识别成功，记录状态，关闭登录页
+    //     User().saveFingerprint(true);
+    //     Navigator.pop(context);
+    //   }else{
+    //     //识别失败，提示设置手势
+    //     onShowDialog(context, '是否设置手势图案快速登录？', (){
+    //       //用户取消设置图案，关闭登录页
+    //       Navigator.pop(context);
+    //     }, ()=>_setGesture(context));
+    //   }
+    // } on PlatformException catch (e) {
+    //   print(e);
+    // }
   }
 
-  void _setGesture(context){
-
-  }
+  void _setGesture(context) {}
 }
